backend/requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
opencv-python
requests
git+https://github.com/ChaoningZhang/MobileSAM.git
python-dotenv
timm
